// client/packages/editor-oss/src/controls/AiWorldController/AiWorldController.types.ts
export type GenerationStep = {
    step: string;
    function: string;
    parameters: any;
    description: string;
};

export enum GENERATION_STEPS_FUNCTIONS {
    ENCHANCE_PROMPT = "enchancePrompt",
    GENERATE_MODEL = "generate3dObject",
    ATTACH_BEHAVIORS = "attachBehaviors",
    MODIFY_MODEL = "modifyModelByCopilot",
}

export enum IMAGE_TYPES {
    CHARACTER = "Character",
    OBJECT = "Object",
    BACKDROP = "Backdrop",
    SKYBOX = "Skybox",
}
export enum TEXTURE_QUALITY {
    STANDARD = "standard",
    DETAILED = "detailed",
}
export enum MODEL_VERSION {
    V_25 = "v2.5-20250123",
    V_20 = "v2.0-20240919",
    V_14 = "v1.4-20240625",
}

export type AISearchTagsResponse = {
    tags: string[];
    width: number;
    height: number;
    followUpMessage: string;
};